changed depencies of core and engine

+++ libs/engine3d/engine/camera_model.h
@@ -43,7 +43,7 @@ public:
     void fillCubeFaces(const std::vector<triangle>* mesh);
     void world_transform(const cv::Mat* matrix, triangle* tri);
     void camera_transform(const cv::Mat* matrix, triangle* tri);
-    cv::Vec3f getCameraVector(const cv::Mat& V_T_C);
+    cv::Vec3d getCameraVector(const cv::Mat& V_T_C);
 
     void resetCameraImage(cv::Mat frame);
     cv::Mat& getCameraImage();
