+++ libs/engine3d/engine/color.h
@@ -11,7 +11,7 @@
 class Color {
 public:
 
-    double intensity(cv::Vec3f light_direction, cv::Mat& normal);
+    double intensity(cv::Vec3d light_direction, cv::Mat& normal);
     cv::Scalar bgr_to_hsl(const cv::Scalar& bgr);
     cv::Scalar hsl_to_bgr(const cv::Scalar& hls);
     cv::Scalar adjust_bgr_intensity(cv::Scalar& base_color, double intensity) ;
